Feature/actualizar carrito (#12)
* Se agregaró la entidad carritoDulce, creamos el repositorio Carritos y agregamos un nuevo metodo al repositorio de dulces, realizamos el provider de carritos_dulces, generamos el caso de uso, realizamos el query movementes_result, creamos el handler update_carrito e implementamos test manuales

* unimos la rama main

* cambios

* cambios

* se realizaron todos los test del repositorio de carritos

* se generó el provider, se realizó el caso de uso donde evalua los moviminetos y devuelve una respuesta a cada movimniento, se realizó un query, se generó una ruta y se cinfiguro el handler. Serealizon test unitarios y manuales además del diagrama uml

* CQ

* CQ

* comentarios Freyman

+++ internal/domain/dto/query/obtener_detalle_dulce.go
@@ -5,7 +5,7 @@ import "Mileyman-API/internal/domain/entities"
 type DetalleDulce struct {
 	ID               uint64                `json:"id"`
 	Peso             int                   `json:"peso"`
-	PrecioUnidad     int                   `json:"precio_unidad"`
+	PrecioUnidad     float64               `json:"precio_unidad"`
 	Disponibles      int                   `json:"disponibles"`
 	Codigo           string                `json:"codigo"`
 	Nombre           string                `json:"nombre"`
